refactor(llm): split two phase pipeline internals
